Fixed: Keep Discord URL helpers client-safe (no node:crypto)
