+++ apps/dashboard/features/products/discord.ts
@@ -193,16 +193,6 @@ export function isDiscordDm(interaction: DiscordInteraction): boolean {
   return !interaction.guild_id;
 }
 
-/** Build a user-install (Add App) URL for Hermies-style DMs. */
-export function buildDiscordUserInstallUrl(clientId: string): string {
-  const params = new URLSearchParams({
-    client_id: clientId,
-    integration_type: "1",
-    scope: "applications.commands",
-  });
-  return `https://discord.com/oauth2/authorize?${params.toString()}`;
-}
-
 /** Follow up on a deferred interaction with the final reply. */
 export async function editDiscordInteractionResponse(
   applicationId: string,
@@ -233,15 +223,7 @@ export async function editDiscordInteractionResponse(
   }
 }
 
-/** Build the OAuth invite URL for a product bot. */
-export function buildDiscordInviteUrl(clientId: string): string {
-  const params = new URLSearchParams({
-    client_id: clientId,
-    scope: "bot applications.commands",
-    permissions: "2147485696", // Send Messages + Use Application Commands
-  });
-  return `https://discord.com/oauth2/authorize?${params.toString()}`;
-}
+export { buildDiscordInviteUrl, buildDiscordUserInstallUrl } from "./discord-urls";
 
 /** Extract a slash-command string option by name. */
 export function getDiscordOptionString(
